==> 1.Feature_Creation.ipynb <==
#import standard data analysis libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

import warnings
warnings.filterwarnings('ignore')

from functions import *

This part of the code contains the functions used later on

#main function dictating which functions to call for calculating player stats
def add_to_df_players(df_players, df_savant, at_bat_ids):
    """
    Enhances the df_players DataFrame by computing and adding various performance statistics 
    based on player event data from df_savant and at_bat_ids.

    This function:
    - Filters out players without an MLB ID.
    - Computes the number of years since a player’s MLB debut.
    - Aggregates plate appearance (PA) and batters faced (BF) counts.
    - Tracks fielding appearances at different positions.
    - Counts the occurrences of various play events (e.g., hits, strikeouts, walks).
    - Computes key batting statistics like batting average, slugging, and on-base percentage.

    Parameters:
    -----------
    df_players : pandas.DataFrame
        DataFrame containing player-level information, including player IDs and debut years.
    df_savant : pandas.DataFrame
        Baseball Savant event-level data, which includes play-by-play information.

    Returns:
    --------
    pandas.DataFrame
        Updated df_players with additional columns
    """
    
    #remove players without an mlb Id
    df_players = df_players[df_players['player_mlb_id'].notna()]

    #calculate years since a player debuted
    df_players['years_after_debut'] = df_players['year'] - df_players['debut'].str[-4:].astype(int)

    #get the counts for the main dataframe
    df_players = compute_player_pa_and_bf(df_players, at_bat_ids, df_savant)

    #calculate where a player played in the field
    df_players = get_fielding_counts(df_players, df_savant)

    #count how many of each stat a player had on the year
    df_players = calculate_all_play_event_counts(df_players,df_savant)

    #calculate average lineup position
    df_players = calculate_lineup_position(df_players, df_savant)
    
    #calculate general batting stats
    df_players = calculate_batting_stats(df_players, df_savant)

    #innings pitched calculation
    df_players = calculate_innings_pitched(df_players, df_savant)

    #average fastball velocity
    df_players = fastball_velocity(df_players, df_savant)
    
    #calculate pitching stats
    df_players = calculate_pitching_stats(df_players)

    #add average batters faced per outing for pitchers
    df_players = calculate_batters_faced_in_game(df_players,df_savant)

    #find inzone + chase rate
    df_players = calculate_zone_chase_pct(df_players, df_savant)
    
    #calculate rbis
    df_players = calculate_rbis(df_players,df_savant)

    #calculate average expected batting avg for batters and pitchers
    df_players = calculate_average_exp_ba(df_players,df_savant,'batter')
    df_players = calculate_average_exp_ba(df_players,df_savant,'pitcher')

    #look at WOBA and xWOBA
    df_players = calculate_woba(df_players, df_savant, 'batter')
    df_players = calculate_woba(df_players, df_savant, 'pitcher')
    df_players = calculate_average_xwoba(df_players, df_savant, 'batter')
    df_players = calculate_average_xwoba(df_players, df_savant, 'pitcher')

    #get player position estimates
    df_players = primary_position(df_players)

    #get percentage of times a pitcher started
    df_players = add_sp_percentage(df_players, df_savant)
    
    #get player age
    df_players['age'] = df_players['year'] - df_players['birthYear']
    
    #years from 26-28 peak age performance (https://tht.fangraphs.com/how-do-baseball-players-age-part-1/)
    df_players['years_before_26'] = (26 - df_players['age']).clip(lower=0)
    df_players['years_after_28'] = (df_players['age'] - 28).clip(lower=0)

    return df_players

Past the functions

#read the csvs as dataframes
df_savant_raw = pd.read_csv('savant_data_2021_2023.csv')
df_people = pd.read_csv('lahman_people.csv')

#create dataframe for player season info
df_players = pd.concat([df_people.assign(year=i) for i in range(2021, 2024)], ignore_index=True)

#clean Savant data
df_savant, at_bat_ids = clean_savant_data(df_savant_raw)

#add columns to df_players based on the Savant data
df_players_full = add_to_df_players(df_players, df_savant, at_bat_ids)

# #get savant data for batting splits
# df_savant_r = df_savant_raw[df_savant_raw['p_throws'] == 'R']
# df_savant_l = df_savant_raw[df_savant_raw['p_throws'] == 'L']

# #make df_players for each batting split
# df_players_r = add_to_df_players(df_players, df_savant_r)
# df_players_l = add_to_df_players(df_players, df_savant_l)

#export csv for use in other .ipynb
df_players_full.to_csv('df_players.csv')
# df_players_r.to_csv('df_players_right.csv')
# df_players_l.to_csv('df_players_left.csv')

for i in df_players_full.columns:
    print(i)
