==> 2.Batter_Features_submission.ipynb <==
#import standard data analysis libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

df_players = pd.read_csv('df_players.csv')

#make only for batting
df_players_batting = df_players[df_players['total_pa'] > 0]

#selected features
current_batter_features = [
    'player_mlb_id',
    'year',
    'age',
    'years_after_28',
    'bats',
    'primary_position',
    'ops',
    'k_rate_batter',
    'bb_rate_batter',
    'batter_avg_exp_ba',
    'babip_batter',
    'fly_balls_ratio_batter',
    'ground_balls_ratio_batter',
    'line_drives_ratio_batter',
    'popups_ratio_batter',
    'hr_fb_pct_batter',
    'avg_lineup_position',
    'batter_avg_xwoba',
    'batter_avg_woba',
    'total_pa'
]

df_players_batting = df_players_batting[current_batter_features]

df_players_batting.head()

#deal with n/a values here and in functions

#set infinite ratios to 1
df_players_batting.replace(np.inf, 1, inplace=True)

#get leaguewide averages for stats
numeric_cols = df_players_batting.select_dtypes(include=np.number).columns.tolist()

stat_averages_league = df_players_batting.groupby('year')[numeric_cols].apply(
    lambda group: group.apply(
        lambda col: np.average(col.dropna(), weights=group.loc[col.dropna().index, 'total_pa']) 
        if col.name not in ['total_pa'] and group.loc[col.dropna().index, 'total_pa'].sum() > 0
        else col.mean(),  # Use col.mean() for total_pa column
        axis=0
    )
)

#get averages by position
numeric_cols = df_players_batting.select_dtypes(include=np.number).columns.tolist()
numeric_cols += ['primary_position']

df_batting_num_cols = df_players_batting[numeric_cols]

#get means
stat_averages_position = df_batting_num_cols.groupby(['year', 'primary_position']).apply(
    lambda group: group.apply(
        lambda col: np.average(col.dropna(), weights=group.loc[col.dropna().index, 'total_pa']) 
        if col.name != 'total_pa' and group.loc[col.dropna().index, 'total_pa'].sum() > 0
        else col.mean(),  # Use col.mean() for total_pa column
        axis=0
    )
).fillna(0).reset_index()

#replace stats for players with unknown positions with league averages
common_cols = stat_averages_league.columns.intersection(stat_averages_position.columns.drop('primary_position'))
for year in stat_averages_position['year'].unique():
    mask = (stat_averages_position['year'] == year) & (stat_averages_position['primary_position'] == 'Unknown')
    stat_averages_position.loc[mask, common_cols] = stat_averages_league.loc[
        stat_averages_league['year'] == year, common_cols
    ].values

#check for nas
df_players_batting.isna().sum()

#these all make sense to set to 0
df_players_batting = df_players_batting.fillna(0)

#put all of player's stats into one row
df_2021 = df_players_batting[df_players_batting['year'] == 2021]
df_2022 = df_players_batting[df_players_batting['year'] == 2022]
df_2023 = df_players_batting[df_players_batting['year'] == 2023]
df_all = df_2021.merge(df_2022, on='player_mlb_id', how='outer', suffixes=("_2ya","_1ya")).drop(columns=['year_2ya','year_1ya'])
df_all = df_all.merge(df_2023, on='player_mlb_id',how='outer')

#remove pitchers due to universal DH adoption (except if they got over 200 PAs in a season (Shohei))
df_all = df_all[((df_all['primary_position_1ya'] != 'p') & (df_all['primary_position_2ya'] != 'p')) | (df_all['total_pa_1ya'] > 200) | (df_all['total_pa_2ya'] > 200)]

#check nulls
df_all.isna().sum()

#fill the non-numerical nulls
df_all['age'] = df_all['age'].fillna(df_all['age_1ya']+1).fillna(df_all['age_2ya'] + 2)  # Fill missing age and adjust

df_all['years_after_28'] = df_all['years_after_28'].fillna(df_all['years_after_28_1ya']+1).fillna(df_all['years_after_28_2ya'] + 2)  # Fill missing years_after_28 and adjust

df_all['primary_position'] = df_all['primary_position'].fillna(df_all['primary_position_1ya']).fillna(df_all['primary_position_2ya']).fillna('Unknown')

df_all['bats'] = df_all['bats'].fillna(df_all['bats_1ya']).fillna(df_all['bats_2ya'])

df_all = df_all.drop(columns=['age_1ya','age_2ya','years_after_28_1ya','years_after_28_2ya','bats_1ya','bats_2ya','primary_position_2ya','primary_position_1ya'])

#everything else can be filled with 0
df_all = df_all.fillna(0)

df_all.columns

#get columns to be averaged
cols_0ya = ['ops', 'k_rate_batter', 
 'bb_rate_batter', 'batter_avg_exp_ba', 'babip_batter', 'batter_avg_xwoba', 'batter_avg_woba',
 'fly_balls_ratio_batter', 'ground_balls_ratio_batter', 
 'line_drives_ratio_batter', 'popups_ratio_batter', 
 'hr_fb_pct_batter', 'avg_lineup_position']


cols_1ya = ['ops_1ya', 'k_rate_batter_1ya', 
            'bb_rate_batter_1ya', 'batter_avg_exp_ba_1ya', 'babip_batter_1ya', 'batter_avg_xwoba_1ya', 'batter_avg_woba_1ya',
            'fly_balls_ratio_batter_1ya', 'ground_balls_ratio_batter_1ya', 
            'line_drives_ratio_batter_1ya', 'popups_ratio_batter_1ya', 
            'hr_fb_pct_batter_1ya', 'avg_lineup_position_1ya']

cols_2ya = ['ops_2ya', 'k_rate_batter_2ya', 
            'bb_rate_batter_2ya', 'batter_avg_exp_ba_2ya', 'babip_batter_2ya', 'batter_avg_xwoba_2ya', 'batter_avg_woba_2ya',
            'fly_balls_ratio_batter_2ya', 'ground_balls_ratio_batter_2ya', 
            'line_drives_ratio_batter_2ya', 'popups_ratio_batter_2ya', 
            'hr_fb_pct_batter_2ya', 'avg_lineup_position_2ya']

pa_0ya = 'total_pa'
pa_1ya = 'total_pa_1ya'
pa_2ya = 'total_pa_2ya'

# Calculate the weighted averages and replace the original columns
for col_0ya, col_1ya, col_2ya in zip(cols_0ya, cols_1ya, cols_2ya):
    new_col_name = col_0ya + '_'
    df_all[new_col_name] = (
        (5 * df_all[col_0ya] * df_all[pa_0ya] + 3 * df_all[col_1ya] * df_all[pa_1ya] + 2 * df_all[col_2ya] * df_all[pa_2ya]) /
        (5 * df_all[pa_0ya] + 3 * df_all[pa_1ya] + 2 * df_all[pa_2ya])
    )

# Drop the original _1ya and _2ya columns
df_all_weighed = df_all.drop(columns=cols_1ya + cols_2ya + cols_0ya)

#do the same for plate appearances
df_all_weighed['total_pa_'] = (5*df_all_weighed['total_pa'] + 3*df_all_weighed['total_pa_1ya'] + 2*df_all_weighed['total_pa_2ya']) / 10

df_all_weighed.drop(columns=['total_pa_1ya','total_pa_2ya','total_pa'],inplace=True)

df_all_weighed.head()

df_all_weighed.select_dtypes(include=np.number).corr()

df_all_weighed.columns

#regress everyone's stats to the mean of their position group by 5 PA (to help with players who had very few plate appearances)
stats_to_regress = ['ops_', 'k_rate_batter_',
       'bb_rate_batter_', 'batter_avg_exp_ba_', 'babip_batter_',
       'batter_avg_xwoba_', 'batter_avg_woba_', 'fly_balls_ratio_batter_',
       'ground_balls_ratio_batter_', 'line_drives_ratio_batter_',
       'popups_ratio_batter_', 'hr_fb_pct_batter_']

    

stat_averages_position_2021_2023 = stat_averages_position.drop(columns='year').reset_index().groupby('primary_position').mean().reset_index()

df_all_weighed_regress = df_all_weighed.merge(stat_averages_position_2021_2023, on='primary_position', suffixes=('', '_incorrect'))

df_all_weighed_regress.isna().sum()

df_all_weighed_regress.columns

for stat in stats_to_regress:
    stat_base = stat.rstrip('_')
    stat_regressed = f"{stat}reg"

    df_all_weighed_regress[stat_regressed] = (
        df_all_weighed_regress[stat] * df_all_weighed_regress["total_pa_"] 
        + df_all_weighed_regress[stat_base] * 5
    ) / (df_all_weighed_regress["total_pa_"] + 5)

    df_all_weighed_regress.drop(columns=[stat, stat_base], inplace=True)

# Remove incorrect columns and the index column
df_all_weighed_regress.drop(columns=df_all_weighed_regress.filter(like="_incorrect").columns, inplace=True)
df_all_weighed_regress.drop(columns="index", inplace=True)


df_all_weighed_regress.to_csv('batter_features_final_model.csv')
