==> 2.Pitcher_Features_submission.ipynb <==
#import standard data analysis libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

df_players = pd.read_csv('df_players.csv')

#make only for batting
df_players_pitching = df_players[df_players['total_bf'] > 0]

#selected features
current_pitcher_features = [
    'player_mlb_id',
    'age',
    'year',
    'primary_position',
    'years_after_28',
    'throws',
    'whip',
    'baa',
    'sp_pct',
    'k_rate_pitcher',
    'pitcher_avg_xwoba',
    'pitcher_avg_woba',
    'bb_rate_pitcher',
    'inplay_ratio_pitcher',
    'pitcher_avg_exp_ba',
    'babip_pitcher',
    'fly_balls_ratio_pitcher',
    'ground_balls_ratio_pitcher',
    'line_drives_ratio_pitcher',
    'popups_ratio_pitcher',
    'hr_fb_pct_pitcher',
    'avg_bf_per_outing',
    'zone_chase_pct',
    'avg_fb_vel',
    'starter',
    'reliever',
    'both_starter_reliever',
    'total_bf'
    #regress stats based on 50 league average plate appearances
]

df_players_pitching = df_players_pitching[current_pitcher_features]

#remove position players who pitched
df_players_pitching = df_players_pitching[df_players_pitching['primary_position'] == 'p']

df_players_pitching.head()

#deal with n/a values here and in functions

#set infinite ratios to 1
df_players_pitching.replace(np.inf, 1, inplace=True)

# Create the 'primary_position' column based on the column with the value of 1
df_players_pitching['primary_position'] = np.select(
    [df_players_pitching['starter'] == 1, 
     df_players_pitching['reliever'] == 1, 
     df_players_pitching['both_starter_reliever'] == 1],
    ['starter', 'reliever', 'both_starter_reliever'],
    default='unknown'  # This will handle the case where none of them are 1, if applicable
)

df_players_pitching = df_players_pitching.drop(columns=['starter','reliever','both_starter_reliever'])

#get leaguewide averages for stats
numeric_cols = df_players_pitching.select_dtypes(include=np.number).columns.tolist()

stat_averages_league = df_players_pitching.groupby('year')[numeric_cols].apply(
    lambda group: group.apply(
        lambda col: np.average(col.dropna(), weights=group.loc[col.dropna().index, 'total_bf']) 
        if col.name not in ['total_bf'] and group.loc[col.dropna().index, 'total_bf'].sum() > 0
        else col.mean(),  # Use col.mean() for total_bf column
        axis=0
    )
)

#get averages by position
numeric_cols = df_players_pitching.select_dtypes(include=np.number).columns.tolist()
numeric_cols += ['primary_position']

df_batting_num_cols = df_players_pitching[numeric_cols]

#get means
stat_averages_position = df_batting_num_cols.groupby(['year', 'primary_position']).apply(
    lambda group: group.apply(
        lambda col: np.average(col.dropna(), weights=group.loc[col.dropna().index, 'total_bf']) 
        if col.name != 'total_bf' and group.loc[col.dropna().index, 'total_bf'].sum() > 0
        else col.mean(),  # Use col.mean() for total_bf column
        axis=0
    )
).fillna(0).reset_index()

#replace stats for players with unknown positions with league averages
common_cols = stat_averages_league.columns.intersection(stat_averages_position.columns.drop('primary_position'))
for year in stat_averages_position['year'].unique():
    mask = (stat_averages_position['year'] == year) & (stat_averages_position['primary_position'] == 'Unknown')
    stat_averages_position.loc[mask, common_cols] = stat_averages_league.loc[
        stat_averages_league['year'] == year, common_cols
    ].values

#fill nas
df_players_pitching['avg_fb_vel'] = df_players_pitching['avg_fb_vel'].fillna((92.756616 + 92.970260) / 2)
df_players_pitching['primary_position'] = df_players_pitching['primary_position'].fillna('both_starter_reliever')
df_players_pitching = df_players_pitching.fillna(0)

#put all of player's stats into one row
df_2021 = df_players_pitching[df_players_pitching['year'] == 2021]
df_2022 = df_players_pitching[df_players_pitching['year'] == 2022]
df_2023 = df_players_pitching[df_players_pitching['year'] == 2023]
df_all = df_2021.merge(df_2022, on='player_mlb_id', how='outer', suffixes=("_2ya","_1ya")).drop(columns=['year_2ya','year_1ya'])
df_all = df_all.merge(df_2023, on='player_mlb_id',how='outer')

#check nulls
df_all.isna().sum()

#fill the non-numerical nulls
df_all['age'] = df_all['age'].fillna(df_all['age_1ya']+1).fillna(df_all['age_2ya'] + 2)  # Fill missing age and adjust

df_all['years_after_28'] = df_all['years_after_28'].fillna(df_all['years_after_28_1ya']+1).fillna(df_all['years_after_28_2ya'] + 2)  # Fill missing years_after_28 and adjust

df_all['primary_position'] = df_all['primary_position_1ya'].fillna(df_all['primary_position_2ya']).fillna('both_starter_reliever')

df_all['throws'] = df_all['throws_1ya'].fillna(df_all['throws_2ya'])

df_all = df_all.drop(columns=['age_1ya','age_2ya','years_after_28_1ya','years_after_28_2ya','throws_1ya','throws_2ya','primary_position_2ya','primary_position_1ya'])

#fill remaining with 0, won't have an impact because remaining nulls are in years where total_bf is 0
df_all = df_all.fillna(0)

df_all.columns

#get columns to be averaged
cols_0ya = ['whip', 'baa', 'sp_pct', 'k_rate_pitcher', 'pitcher_avg_xwoba', 'pitcher_avg_woba',
 'bb_rate_pitcher', 'inplay_ratio_pitcher', 'pitcher_avg_exp_ba', 'babip_pitcher',
 'fly_balls_ratio_pitcher', 'ground_balls_ratio_pitcher', 'line_drives_ratio_pitcher',
 'popups_ratio_pitcher', 'hr_fb_pct_pitcher', 'avg_bf_per_outing', 'zone_chase_pct', 'avg_fb_vel']


cols_1ya = ['whip_1ya', 'baa_1ya', 'sp_pct_1ya', 'k_rate_pitcher_1ya','pitcher_avg_xwoba_1ya','pitcher_avg_woba_1ya',
         'bb_rate_pitcher_1ya', 'inplay_ratio_pitcher_1ya',
         'pitcher_avg_exp_ba_1ya', 'babip_pitcher_1ya',
         'fly_balls_ratio_pitcher_1ya', 'ground_balls_ratio_pitcher_1ya', 'line_drives_ratio_pitcher_1ya',
         'popups_ratio_pitcher_1ya', 'hr_fb_pct_pitcher_1ya',
         'avg_bf_per_outing_1ya', 'zone_chase_pct_1ya', 'avg_fb_vel_1ya']

cols_2ya = ['whip_2ya', 'baa_2ya', 'sp_pct_2ya', 'k_rate_pitcher_2ya','pitcher_avg_xwoba_2ya','pitcher_avg_woba_2ya',
           'bb_rate_pitcher_2ya', 'inplay_ratio_pitcher_2ya',
           'pitcher_avg_exp_ba_2ya', 'babip_pitcher_2ya',
           'fly_balls_ratio_pitcher_2ya', 'ground_balls_ratio_pitcher_2ya', 'line_drives_ratio_pitcher_2ya',
           'popups_ratio_pitcher_2ya', 'hr_fb_pct_pitcher_2ya',
           'avg_bf_per_outing_2ya', 'zone_chase_pct_2ya', 'avg_fb_vel_2ya']

pa_0ya = 'total_bf'
pa_1ya = 'total_bf_1ya'
pa_2ya = 'total_bf_2ya'

# Calculate the weighted averages and replace the original columns
for col_0ya, col_1ya, col_2ya in zip(cols_0ya, cols_1ya, cols_2ya):
    new_col_name = col_0ya + '_'
    df_all[new_col_name] = (
        (5 * df_all[col_0ya] * df_all[pa_0ya] + 3 * df_all[col_1ya] * df_all[pa_1ya] + 2 * df_all[col_2ya] * df_all[pa_2ya]) /
        (5 * df_all[pa_0ya] + 3 * df_all[pa_1ya] + 2 * df_all[pa_2ya])
    )

# Drop the original _1ya and _2ya columns
df_all_weighed = df_all.drop(columns=cols_1ya + cols_2ya + cols_0ya)

#do the same for plate appearances
df_all_weighed['total_bf_'] = (5*df_all_weighed['total_bf'] + 3*df_all_weighed['total_bf_1ya'] + 2*df_all_weighed['total_bf_2ya']) / 10

df_all_weighed.drop(columns=['total_bf','total_bf_1ya','total_bf_2ya'],inplace=True)

df_all_weighed.head()

df_all_weighed.select_dtypes(include=np.number).corr()

df_all_weighed.columns

#regress everyone's stats to the mean of their position group by 5 PA (to help with players who had very few plate appearances)
stats_to_regress = ['whip_', 'baa_', 'k_rate_pitcher_', 'bb_rate_pitcher_',
       'inplay_ratio_pitcher_', 'pitcher_avg_exp_ba_', 'babip_pitcher_',
       'fly_balls_ratio_pitcher_', 'ground_balls_ratio_pitcher_', 'line_drives_ratio_pitcher_',
       'popups_ratio_pitcher_', 'hr_fb_pct_pitcher_', 'avg_bf_per_outing_','pitcher_avg_xwoba_','pitcher_avg_woba_',
       'zone_chase_pct_', 'avg_fb_vel_']

stat_averages_position_2021_2023 = stat_averages_position.drop(columns='year').reset_index().groupby('primary_position').mean().reset_index()

df_all_weighed_regress = df_all_weighed.merge(stat_averages_position_2021_2023, on='primary_position', suffixes=('', '_incorrect'))

for stat in stats_to_regress:
    stat_base = stat.rstrip('_')
    stat_regressed = f"{stat}reg"

    df_all_weighed_regress[stat_regressed] = (
        df_all_weighed_regress[stat] * df_all_weighed_regress["total_bf_"] 
        + df_all_weighed_regress[stat_base] * 5
    ) / (df_all_weighed_regress["total_bf_"] + 5)

    df_all_weighed_regress.drop(columns=[stat, stat_base], inplace=True)

# Remove incorrect columns and the index column
df_all_weighed_regress.drop(columns=df_all_weighed_regress.filter(like="_incorrect").columns, inplace=True)
df_all_weighed_regress.drop(columns="index", inplace=True)

df_all_weighed_regress

df_all_weighed_regress.to_csv('pitcher_features_final_model.csv')
